2024-12-simple-rag/requirements.txt
elasticsearch==8.15.0
python-dotenv==1.0.1
streamlit==1.39.0
langchain-cohere==0.3.1
